+++ app/src/main/kotlin/me/xizzhu/android/joshua/reading/toolbar/ReadingToolbarInteractor.kt
@@ -63,7 +63,5 @@ class ReadingToolbarInteractor(private val bibleReadingManager: BibleReadingMana
 
     fun bookShortNames(): Flow<ViewData<List<String>>> = currentTranslation()
             .filterOnSuccess()
-            .map { bibleReadingManager.readBookShortNames(it) }
-            .filter { it.isNotEmpty() }
-            .toViewData()
+            .map { ViewData.success(bibleReadingManager.readBookShortNames(it)) }
 }