+++ docs/CHANGELOG.md
@@ -1,6 +1,11 @@
 CHANGELOG
 ---------
 
+#### Next Release
+- Properly handles case when translation downloading is cancelled
+  - As a result, reverted "Only emit if book names and book short names are loaded"
+- Throw exception when missing translation list
+
 #### v0.12.5 (2019-12-07)
 - Do nothing if it's already downloading or removing translations
 - Fixed searching with single quote